sim_anomalies/requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
requests==2.31.0
PyYAML==6.0.1
docker==6.1.3
influxdb-client==1.38.0
prometheus-api-client==0.5.4
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
pyshark==0.6